build: devDependencies and tsup (#29)
