+++ .gitignore
@@ -23,4 +23,4 @@ dist-ssr
 *.sln
 *.sw?
 .cursor/*
-memory-bank/*
+.memory/*
