+++ package.json
@@ -48,7 +48,6 @@
   },
   "homepage": "https://www.npmjs.com/package/zusound",
   "dependencies": {
-    "tsup": "^8.4.0",
     "zustand": "^5.0.3"
   },
   "peerDependencies": {
@@ -76,6 +75,7 @@
     "react-syntax-highlighter": "^15.6.1",
     "rimraf": "^6.0.1",
     "tsc-alias": "^1.8.13",
+    "tsup": "^8.4.0",
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.24.1",
     "vite": "^6.2.0",
