+++ packages/middleware/zusound.ts
@@ -1,14 +1,25 @@
 import type { StoreMutatorIdentifier, StateCreator } from 'zustand'
-import type { TraceOptions, TraceData } from '../core' // Core trace functionality
+import type { TraceOptions, TraceData, ZusoundTraceEventDetail } from '../core' // Import ZusoundTraceEventDetail from core
 import { trace } from '../core' // Core trace functionality
 import type { Zusound, ZusoundOptions } from './types' // Middleware-specific types
 import { isProduction } from './utils' // Utility functions
-import { sonifyChanges } from '../sonification' // Sonification function
-import {
-  ensureVisualizerReady,
-  showPersistentVisualizer as showVisualizerUI, // Rename import for clarity
-  // hidePersistentVisualizer as hideVisualizerUI, // This function is not used locally
-} from '../visualizer' // Visualizer UI control
+import { initSonificationListener } from '../sonification' // Import for initialization but not direct calls
+
+// Define the custom event type using the detail from core
+export type ZusoundTraceEvent<T = unknown> = CustomEvent<ZusoundTraceEventDetail<T>>
+
+/**
+ * Dispatches a trace event containing the trace data.
+ * Visualizer and sonification components can listen for this event.
+ */
+function dispatchTraceEvent<T>(traceData: TraceData<T>): void {
+  if (typeof window !== 'undefined') {
+    const event = new CustomEvent<ZusoundTraceEventDetail<T>>('zusound:trace', {
+      detail: { traceData },
+    })
+    window.dispatchEvent(event)
+  }
+}
 
 /**
  * zusound middleware for Zustand
@@ -26,9 +37,9 @@ export const zusound: Zusound = <
     enabled = !isProduction(),
     logDiffs = false,
     allowInProduction = false,
-    onTrace: userOnTrace, // User-defined callback after sonification
+    onTrace: userOnTrace, // User-defined callback *after* event dispatch
     diffFn, // Custom diff function
-    persistVisualizer = false, // Option to show persistent UI on init
+    initSonification = true, // Default to true for backward compatibility
     ...restOptions // Remaining options passed down to core trace
   } = options
 
@@ -40,26 +51,22 @@ export const zusound: Zusound = <
     return initializer as StateCreator<T, Mps, [...Mcs]>
   }
 
-  // Initialize visualizer and show persistent UI if requested (runs once)
-  if (typeof window !== 'undefined') {
+  // Initialize sonification listener if requested (runs once)
+  if (typeof window !== 'undefined' && initSonification) {
     // Use setTimeout to ensure DOM is ready and avoid blocking initial render
     setTimeout(() => {
       try {
-        ensureVisualizerReady() // Ensures the visualizer singleton is initialized
-        if (persistVisualizer) {
-          showVisualizerUI() // Show the persistent UI element
-        }
+        initSonificationListener() // Initialize sonification listener
       } catch (error) {
-        console.error('Error initializing visualizer:', error)
+        console.error('Error initializing sonification:', error)
       }
     }, 0)
   }
 
   // --- Custom trace processor (internal onTrace for core) ---
   const coreOnTrace = (traceData: TraceData<T>) => {
-    // 1. Perform Sonification
-    // SonifyChanges now dispatches events internally, visualizer listens separately.
-    sonifyChanges(traceData.diff, traceData.duration)
+    // 1. Dispatch the trace event for sonification and visualization
+    dispatchTraceEvent(traceData)
 
     // 2. Call user's onTrace callback if provided
     if (userOnTrace) {
@@ -75,7 +82,7 @@ export const zusound: Zusound = <
   const traceOptions: TraceOptions<T> = {
     ...restOptions, // Pass down any remaining compatible options
     diffFn, // Pass down custom diff function if provided
-    onTrace: coreOnTrace, // Use the internal handler that triggers sonification & user callback
+    onTrace: coreOnTrace, // Use the internal handler that dispatches events & calls user callback
   }
 
   // --- Setup Logging ---
@@ -91,7 +98,7 @@ export const zusound: Zusound = <
         // Push a copy to avoid mutation issues if traceData is modified later
         window['__zusound_logger__'].push({ ...traceData })
       }
-      // Call the original handler (which does sonification + user callback)
+      // Call the original handler (which dispatches events + user callback)
       originalCoreOnTrace(traceData)
     }
   }
