+++ packages/visualizer/src/config.ts
@@ -1,0 +1,10 @@
+// Visualizer configuration constants
+
+/** Size of the square canvas for the visualizer (width & height) */
+export const VISUALIZER_SIZE = 64
+
+/** Maximum number of concurrent visualization events to track and display */
+export const MAX_VISIBLE_EVENTS = 10
+
+/** Duration in milliseconds each visualization event remains visible */
+export const EVENT_LIFETIME_MS = 1000
